+++ visualization/app/codeCharta/core/core.module.ts
@@ -7,6 +7,7 @@ import "./url/url.module";
 import "./treemap/treemap.module";
 import "./scenario/scenario.module";
 import "./tooltip/tooltip.module";
+import "./download/download.module";
 import "./statistic/statistic.module";
 
 import angular from "angular";
@@ -21,6 +22,7 @@ angular.module(
         "app.codeCharta.core.treemap",
         "app.codeCharta.core.scenario",
         "app.codeCharta.core.tooltip",
+        "app.codeCharta.core.download",
         "app.codeCharta.core.statistic"
     ]
 );