+++ visualization/app/codeCharta/ui/codeMap/codeMap.render.service.ts
@@ -97,6 +97,7 @@ export class CodeMapRenderService implements SettingsServiceSubscriber, CodeMapM
             margin: s.margin,
             invertHeight: s.invertHeight,
             visibleEdges: visibleEdges,
+            blacklist: s.blacklist
         };
 
         let nodes: node[] = this.collectNodesToArray(