+++ visualization/app/codeCharta/ui/codeMap/codeMap.ts
@@ -7,20 +7,22 @@ import {codeMapComponent} from "./codeMap.component";
 import {CodeMapMouseEventService} from "./codeMap.mouseEvent.service";
 import {CodeMapRenderService} from "./codeMap.render.service";
 import {CodeMapActionsService} from "./codeMap.actions.service";
+import {CodeMapUtilService} from "./codeMap.util.service";
 
 angular.module("app.codeCharta.ui.codeMap",["app.codeCharta.ui.codeMap.threeViewer", "app.codeCharta.core"])
 .component(
     codeMapComponent.selector,
     codeMapComponent
-)
-.service(
+).service(
     CodeMapRenderService.SELECTOR,
     CodeMapRenderService
-)
-.service(
+).service(
     CodeMapMouseEventService.SELECTOR,
     CodeMapMouseEventService
 ).service(
     CodeMapActionsService.SELECTOR,
     CodeMapActionsService
+).service(
+    CodeMapUtilService.SELECTOR,
+    CodeMapUtilService
 );